laravel 8 support

+++ src/resources/migrations/2014_02_04_000000_create_country_state_city_table.php
@@ -15,7 +15,7 @@ class CreateCountryStateCityTable extends Migration
     public function up()
     {
         Schema::create('countries', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name',255);
             $table->enum('status', ['active', 'inactive'])->default('active');
             $table->timestamp('created_at')->useCurrent();
@@ -24,8 +24,8 @@ class CreateCountryStateCityTable extends Migration
         });
 
         Schema::create('states', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('country_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('country_id');
             $table->string('name',255);
             $table->enum('status', ['active', 'inactive'])->default('active');
             $table->timestamp('created_at')->useCurrent();
@@ -37,8 +37,8 @@ class CreateCountryStateCityTable extends Migration
         });
 
         Schema::create('cities', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('state_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('state_id');
             $table->string('name',255);
             $table->enum('status', ['active', 'inactive'])->default('active');
             $table->timestamp('created_at')->useCurrent();
